+++ Vec3.h
@@ -6,28 +6,28 @@
 
 class Vec3 {
 public:
-    float v[3];
+    double v[3];
 
     Vec3() {}
 
-    Vec3(float v0, float v1, float v2) {
+    Vec3(double v0, double v1, double v2) {
         v[0] = v0;
         v[1] = v1;
         v[2] = v2;
     }
 
-    float x() const { return v[0]; }
-    float y() const { return v[1]; }
-    float z() const { return v[2]; }
+    double x() const { return v[0]; }
+    double y() const { return v[1]; }
+    double z() const { return v[2]; }
 
-    float r() const { return v[0]; }
-    float g() const { return v[1]; }
-    float b() const { return v[2]; }
+    double r() const { return v[0]; }
+    double g() const { return v[1]; }
+    double b() const { return v[2]; }
 
     const Vec3& operator+() const { return *this; }
     Vec3 operator-() const { return Vec3(-v[0], -v[1], -v[2]); }
-    float operator[](int i) const { return v[i]; }
-    float& operator[](int i) { return v[i]; }
+    double operator[](int i) const { return v[i]; }
+    double& operator[](int i) { return v[i]; }
 
     Vec3& operator+=(const Vec3& v2) {
         v[0] += v2.v[0];
@@ -45,14 +45,14 @@ public:
 
     Vec3& operator*=(const Vec3& v2);
     Vec3& operator/=(const Vec3& v2);
-    Vec3& operator*=(const float t);
-    Vec3& operator/=(const float t);
+    Vec3& operator*=(const double t);
+    Vec3& operator/=(const double t);
     Vec3& operator=(const Vec3& other);
 
 
-    float length() const;
-    float length_squared() const;
-    Vec3 pointOnVector(float t) const;
+    double length() const;
+    double length_squared() const;
+    Vec3 pointOnVector(double t) const;
 
     //make current vector a unit vector
     void normalise();
@@ -82,15 +82,15 @@ inline Vec3& Vec3::operator/=(const Vec3& v2) {
     return *this;
 }
 
-inline Vec3& Vec3::operator*=(const float t) {
+inline Vec3& Vec3::operator*=(const double t) {
     v[0] *= t;
     v[1] *= t;
     v[2] *= t;
     return *this;
 }
 
-inline Vec3& Vec3::operator/=(const float t) {
-    float k = 1.0f / t;
+inline Vec3& Vec3::operator/=(const double t) {
+    double k = 1.0f / t;
 
     v[0] *= k;
     v[1] *= k;
@@ -107,20 +107,20 @@ inline Vec3& Vec3::operator=(const Vec3& other) {
     return *this;
 }
 
-inline float Vec3::length() const {
+inline double Vec3::length() const {
     return std::sqrt(v[0] * v[0] + v[1] * v[1] + v[2] * v[2]);
 }
 
-inline float Vec3::length_squared() const {
+inline double Vec3::length_squared() const {
     return v[0] * v[0] + v[1] * v[1] + v[2] * v[2];
 }
 
-inline Vec3 Vec3::pointOnVector(float t) const {
+inline Vec3 Vec3::pointOnVector(double t) const {
     return Vec3(v[0] * t, v[1] * t, v[2] * t);
 }
 
 inline void Vec3::normalise() {
-    float len = length();
+    double len = length();
 
     v[0] /= len;
     v[1] /= len;
@@ -128,7 +128,7 @@ inline void Vec3::normalise() {
 }
 
 inline Vec3 Vec3::make_normalised() const {
-    float len = length();
+    double len = length();
 
     return Vec3(v[0] / len, v[1] / len, v[2] / len);
 }
@@ -159,23 +159,23 @@ inline Vec3 operator/(const Vec3& v1, const Vec3& v2) {
     return Vec3(v1.v[0] / v2.v[0], v1.v[1] / v2.v[1], v1.v[2] / v2.v[2]);
 }
 
-inline Vec3 operator*(float t, const Vec3& v) {
+inline Vec3 operator*(double t, const Vec3& v) {
     return Vec3(t * v.v[0], t * v.v[1], t * v.v[2]);
 }
 
-inline Vec3 operator*(const Vec3& v, float t) {
+inline Vec3 operator*(const Vec3& v, double t) {
     return Vec3(t * v.v[0], t * v.v[1], t * v.v[2]);
 }
 
-inline Vec3 operator/(float t, const Vec3& v) {
+inline Vec3 operator/(double t, const Vec3& v) {
     return Vec3(t / v.v[0], t / v.v[1], t / v.v[2]);
 }
 
-inline Vec3 operator/(const Vec3& v, float t) {
+inline Vec3 operator/(const Vec3& v, double t) {
     return Vec3(v.v[0] / t, v.v[1] / t, v.v[2] / t);
 }
 
-inline float dot(const Vec3& v1, const Vec3& v2) {
+inline double dot(const Vec3& v1, const Vec3& v2) {
     return v1.v[0] * v2.v[0] + v1.v[1] * v2.v[1] + v1.v[2] * v2.v[2];
 }
